% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbiome-reader.R
\name{import.eupath}
\alias{import.eupath}
\title{Function to read input files for MicrobiomeDB and returns a MicrobiomeData
  object.}
\usage{
import.eupath(taxa_abundance_path, sample_path, aggregate_by = "Phylum",
  use_relative_abundance = T)
}
\arguments{
\item{taxa_abundance_path}{The path to the taxa abundance file.}

\item{sample_path}{The path to the sample characteristics file.}

\item{aggregate_by}{Taxonomic level to aggregate your data (Phylum, Class, Order,
Family, Genus, Species). Defaults to Phylum.}

\item{use_relative_abundance}{If it's true the relative abundance will be used to manipulate
the data, otherwise the absolute abundance will be used. Defaults to TRUE.}
}
\value{
Returns a MicrobiomeData object.
}
\description{
Function to read input files for MicrobiomeDB and returns a MicrobiomeData
  object.
}
\keyword{microbiome}
\keyword{microbiomedb}
