% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_age_dec.R
\name{jd_age_dec}
\alias{jd_age_dec}
\title{Convert age (numeric) to categorical variable (by decades)}
\usage{
jd_age_dec(df, agevar)
}
\arguments{
\item{df}{A dataframe}

\item{agevar}{a character string of the age variable in the dataframe}
}
\value{
A variable of age classified by decades
}
\description{
Convert age (numeric) to categorical variable (by decades)
}
