% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_alpha_stat.R
\name{jd_alpha_stat}
\alias{jd_alpha_stat}
\title{Calculates alpha-diversity statistics}
\usage{
jd_alpha_stat(ps, var, measure)
}
\arguments{
\item{ps}{A phyloseq object}

\item{var}{A variable to associate with the chosen alpha-diversity index}

\item{measure}{An alpha-diversity index}
}
\description{
Calculates alpha-diversity statistics
}
