% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_annotate.R
\name{jd_annotate}
\alias{jd_annotate}
\title{Extract taxa names}
\usage{
jd_annotate(ps, df)
}
\arguments{
\item{ps}{A phyloseq object from where the taxa names will be extracted}

\item{df}{A dataframe, whose rownames are taxa codes}
}
\value{
data frame containing taxonomy names
}
\description{
Extract taxa names
}
