% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_bmi_cat.R
\name{jd_bmi_cat}
\alias{jd_bmi_cat}
\title{Convert BMI (numeric) to category}
\usage{
jd_bmi_cat(df, bmivar)
}
\arguments{
\item{df}{A dataframe}

\item{bmivar}{a character string of the BMI variable in the dataframe}
}
\value{
A variable of BMI classified by BMI categories based on MOH Malaysia
}
\description{
Convert BMI (numeric) to category
}
