% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_bp_cat.R
\name{jd_bp_cat}
\alias{jd_bp_cat}
\title{Convert BP systolic and diastolic (numeric) to category}
\usage{
jd_bp_cat(df)
}
\arguments{
\item{df}{A dataframe, with factor containing BMI variable named as 'bp_dia' and 'bp_sys'}
}
\value{
A variable of BP classified based on the American Heart Association (https://www.heart.org/en/health-topics/high-blood-pressure/understanding-blood-pressure-readings)
}
\description{
Convert BP systolic and diastolic (numeric) to category
}
