% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_dada2.R
\name{jd_dada2}
\alias{jd_dada2}
\title{Run DADA2 ASV inference from primer- and adapter-removed fastq files}
\usage{
jd_dada2(seqlength_min = 400, seqlength_max = 470, truncLen = c(0, 0))
}
\arguments{
\item{seqlength_min}{minimum sequence length to keep, default to 400 for V3-V4 region}

\item{seqlength_max}{maximum sequence length to keep, default to 470 for V3-V4 region}

\item{trunclen}{= dada2 input determining the length of sequence to keep after quality control. Default to no trimming c(0,0)}
}
\value{
csv files of filtered reads summary, ASV sequence table in RDS format, PAPRICA-friendly redundant fasta sequence files, saved R environment
}
\description{
Function must be run on folders containing the fastq files to be inferred to ASV. This function will output several files,
including the list of filtered reads, sequence table file in RDS form (chimera not removed!), redundant fasta files for PAPRICA input,
and save the whole environment (.R file)
}
\details{
It is important to take note that different sample batch should be run separately. Recommended to keep each batch on separate folder
and run this function on a folder loop. For example:

for (directory in 1:length(list.dirs(recursive = F))) {
path = getwd()
dire = list.dirs(recursive = F)
setwd(dire\link{directory})
jd_dada2()
setwd(path)
}
}
