% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_deseq.R
\name{jd_deseq}
\alias{jd_deseq}
\title{Calculates differential abundance statistics based on DESeq2}
\usage{
jd_deseq(ps, var, threshold = 1.5)
}
\arguments{
\item{ps}{A phyloseq object}

\item{var}{A variable to analyse}

\item{threshold}{A number to filter features with less than the chosen number log-fold changes. Defaults to 1.5}
}
\description{
Calculates differential abundance statistics based on DESeq2
}
