% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_drm.R
\name{jd_drm}
\alias{jd_drm}
\title{Calculates and plots Dirichlet Multinomial Mixture model of the dataset}
\usage{
jd_drm(ps, nvar, var)
}
\arguments{
\item{ps}{A phyloseq object -- raw read counts}

\item{nvar}{Number of variables to optimise number of enterotypes}

\item{var}{A variable to classify the detected enterotypes}
}
\value{
A list of the detected enterotypes composition, metadata on the participants' enterotype, and ggplot2 object
}
\description{
Calculates and plots Dirichlet Multinomial Mixture model of the dataset
}
