% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_permanova.R
\name{jd_permanova}
\alias{jd_permanova}
\title{Calculates PERMANOVA}
\usage{
jd_permanova(ps, var, strata = NA, mtd = "euclidean")
}
\arguments{
\item{ps}{A phyloseq object -- raw reads}

\item{var}{A variable(s) to test for PERMANOVA}

\item{strata}{A variable to stratify the PERMANOVA calculation. Defaults to NA}
}
\value{
A list of PERMANOVA output
}
\description{
Calculates PERMANOVA
}
