% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_plot_alpha2.R
\name{jd_plot_alpha2}
\alias{jd_plot_alpha2}
\title{Plots alpha-diversity with statistics}
\usage{
jd_plot_alpha2(ps, var, measure = "Shannon", stat = "kruskal.test")
}
\arguments{
\item{ps}{A phyloseq object}

\item{var}{A variable(s) to classify visualisation}

\item{measure}{Alpha-diversity measures to visualise. Singular value. Defaults to Shannon}

\item{stat}{Statistical method to conduct. Defaults to Kruskal-Wallis}
}
\description{
Plots alpha-diversity with statistics
}
