% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_plot_demo.R
\name{jd_plot_demo}
\alias{jd_plot_demo}
\title{Plot demographic distribution}
\usage{
jd_plot_demo(ps, var, fill)
}
\arguments{
\item{ps}{A phyloseq object}

\item{var}{A variable to plot}

\item{fill}{A variable to classify}
}
\value{
a ggplot2 object
}
\description{
This functions plots the demographic distribution of the chosen variable(s)
}
