% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_plot_ord.R
\name{jd_plot_ord}
\alias{jd_plot_ord}
\title{Ordinates metagenomic data based on variable(s)}
\usage{
jd_plot_ord(ps, var, method = "PCoA", distance = "euclidean")
}
\arguments{
\item{ps}{A phyloseq object}

\item{var}{a character string of the chosen variable. Must be one of the variables in sample_data(ps). Support multiple input variables}

\item{method}{Ordination plot used. Defaults to PCoA}

\item{distance}{A distance matrix for the ordination. Defaults to Euclidean}
}
\description{
Ordinates metagenomic data based on variable(s)
}
