% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_rds.R
\name{jd_rds}
\alias{jd_rds}
\title{Merge ASV sequence tables (RDS files) from DADA2 and annotate taxonomy using SILVA}
\usage{
jd_rds(
  seqlength_min = 400,
  seqlength_max = 470,
  refpath = "E:/ubuntu/virtual_shared/silva_nr99_v138_wSpecies_train_set.fa.gz"
)
}
\arguments{
\item{seqlength_min}{minimum sequence length to keep, default to 400 for V3-V4 region}

\item{seqlength_max}{maximum sequence length to keep, default to 470 for V3-V4 region}

\item{refpath}{path to SILVA reference database}
}
\value{
csv files of abundance and taxonomy table
}
\description{
Function must be run on main folder containing all seqtab.RDS files to be annotated (recursive folders supported).
This function will output abundance and taxonomy tables in the working directory.
}
