% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_spiec.R
\name{jd_spiec}
\alias{jd_spiec}
\title{Spiec Easi network analysis}
\usage{
jd_spiec(ps, var)
}
\arguments{
\item{ps}{A Phyloseq object}

\item{var}{A variable(s) to test for}
}
\value{
list of Spiec Easi object and ggplot2 object
}
\description{
Spiec Easi network analysis
}
