% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_splsda.R
\name{jd_splsda}
\alias{jd_splsda}
\title{Runs PLSDA and SPLSDA for the visualisation, feature-selection and validation of a microbiome dataset}
\usage{
jd_splsda(ps, var, valmodel = "all", no.comp, no.repeat)
}
\arguments{
\item{ps}{A phyloseq object -- raw read count}

\item{var}{A variable(s) for feature selection}

\item{valmodel}{A validation model. To be improved. Defaults to 'all'}

\item{no.comp}{Component numbers for PLSDA and SPLSDA to analyse}

\item{no.repeat}{No. of repeat to be performed by the model}
}
\value{
output files in a new folder listing the compiled details and plots of SPLSDA and SPLSA along with the feature selection plot
}
\description{
Runs PLSDA and SPLSDA for the visualisation, feature-selection and validation of a microbiome dataset
}
