% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_vst.R
\name{jd_vst}
\alias{jd_vst}
\title{Transform raw read counts using VariableStabilizingTransformation in DESeq2}
\usage{
jd_vst(ps)
}
\arguments{
\item{ps}{A phyloseq object}
}
\value{
A phyloseq object with a VST-transformed reads
}
\description{
Transform raw read counts using VariableStabilizingTransformation in DESeq2
}
