% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_write_relbun.R
\name{jd_write_relbun}
\alias{jd_write_relbun}
\title{Calculates the relative abundance value of taxa based on the chosen variable}
\usage{
jd_write_relbun(ps, var)
}
\arguments{
\item{ps}{A phyloseq object -- raw reads}

\item{var}{Variable(s) to calculate relative abundance data}
}
\value{
CSV files containing the relative abundance of each taxa per variable
}
\description{
Calculates the relative abundance value of taxa based on the chosen variable
}
