% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_age_dec.R
\name{jd_age_dec}
\alias{jd_age_dec}
\title{Convert age (numeric) to categorical variable (by decades)}
\usage{
jd_age_dec(df)
}
\arguments{
\item{df}{A dataframe, with factor containing age variable named as 'age'}
}
\value{
A variable of age classified by decades
}
\description{
Convert age (numeric) to categorical variable (by decades)
}
