% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_bmi_cat.R
\name{jd_bmi_cat}
\alias{jd_bmi_cat}
\title{Convert BMI (numeric) to category}
\usage{
jd_bmi_cat(df)
}
\arguments{
\item{df}{A dataframe, with factor containing BMI variable named as 'bmi'}
}
\value{
A variable of BMI classified by BMI categories based on MOH Malaysia
}
\description{
Convert BMI (numeric) to category
}
