% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_clr.R
\name{jd_clr}
\alias{jd_clr}
\title{Transform raw read counts using centered-log transformation approach}
\usage{
jd_clr(ps)
}
\arguments{
\item{ps}{a phyloseq object with raw read counts}
}
\value{
a phyloseq object with CLR-transformed read counts
}
\description{
Transform raw read counts using centered-log transformation approach
}
