% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_plot_alpha.R
\name{jd_plot_alpha}
\alias{jd_plot_alpha}
\title{Plot alpha diversity based on selected variables}
\usage{
jd_plot_alpha(ps, var, measures = c("Chao1", "Shannon"))
}
\arguments{
\item{ps}{A phyloseq object}

\item{var}{Variable(s) to organise visualisation}

\item{measures}{Alpha-diversity measures (for jd_plot_alpha, defaults to Chao1 and Shannon. For jd_plot_alpha_stat, defaults to Shannon. Singular value only)}

\item{stat}{Statistical method to conduct. Defaults to Kruskal-Wallis}
}
\description{
Plot alpha diversity based on selected variables
}
