% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_plot_ord.R
\name{jd_plot_ord}
\alias{jd_plot_ord}
\title{Ordinates metagenomic data based on variable(s)}
\usage{
jd_plot_ord(ps, var, method = "PCoA", distance = "euclidean")
}
\arguments{
\item{ps}{A phyloseq object}

\item{var}{A variable to ordinate}

\item{method}{Ordination plot used}

\item{distance}{A distance matrix for the ordination}
}
\description{
Ordinates metagenomic data based on variable(s)
}
