% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_plot_rel.R
\name{jd_plot_rel}
\alias{jd_plot_rel}
\title{Plots the relative abundance of a microbiome dataset}
\usage{
jd_plot_rel(ps, var, taxrank)
}
\arguments{
\item{ps}{A phyloseq object -- raw read counts}

\item{var}{A variable(s) to organise visualisation}

\item{taxrank}{A taxa level to visualise}
}
\value{
A ggplot2 object
}
\description{
Plots the relative abundance of a microbiome dataset
}
